Squashed commit of the following:
commit 3dcd796630137fd73eda19087f23a58f211b24a7
Author: Karan <karan.batavia@privado.ai>
Date:   Wed Jan 25 00:05:34 2023 +0530

    added comments for subscan times

commit 0aa2f2844f5579aa5ae40ddbcbe8c218eddf1810
Merge: 529508f 35eae31
Author: Karan <karan.batavia@privado.ai>
Date:   Wed Jan 25 00:00:58 2023 +0530

    Merge branch 'dev' of https://github.com/ankit-privado/standalone-monitoring-stability into dev

commit 529508fc4dc46257b226baed338efb302bbe2789
Author: Karan <karan.batavia@privado.ai>
Date:   Wed Jan 25 00:00:14 2023 +0530

    map generation for subscan times

commit 35eae31b16728ffc1b31ed34e193977defd538ea
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Tue Jan 24 19:43:58 2023 +0530

    path analysis fix

commit b13e6002c0555c44c9dda707a280d2c18cb8f621
Merge: cb05986 eddf9c9
Author: Karan Batavia <118820668+karan-batavia@users.noreply.github.com>
Date:   Tue Jan 24 15:43:15 2023 +0530

    Merge pull request #14 from karan-batavia/dev

    Dev

commit eddf9c9ad95c957ec404cabaae8f4a7f2d976305
Author: Karan <karan.batavia@privado.ai>
Date:   Tue Jan 24 15:39:29 2023 +0530

    command output for binary to file added

commit dc1bca9f8f71fc174db0f0df013acfa4df9c07b5
Author: Karan <karan.batavia@privado.ai>
Date:   Tue Jan 24 15:38:47 2023 +0530

    command output for binary to file added

commit b645bb38d90db4c40fdfd6cd21122649e80e5f56
Author: Karan <karan.batavia@privado.ai>
Date:   Tue Jan 24 15:38:09 2023 +0530

    command output for binary to file added

commit cb05986217891c4eac218756d77b6489fb08f139
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Tue Jan 24 15:28:26 2023 +0530

    Added Privado file compare support

commit 9ea9d83e7f7a940d4ee3b004b134283eab0908d3
Merge: feaa3df 720b1d9
Author: Ankit Kumar <118803988+ankit-privado@users.noreply.github.com>
Date:   Mon Jan 23 14:59:38 2023 +0530

    Merge pull request #13 from ankit-privado/aws-refactoring

    Process Error Fix

commit 720b1d93acd1ec3777260659a53bd42af7da5f79
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Mon Jan 23 14:42:15 2023 +0530

    Process Error Fix

commit feaa3df88c6974e36b7c4bcf6589679d10a70fba
Merge: 4222b7b 22596a8
Author: Ankit Kumar <118803988+ankit-privado@users.noreply.github.com>
Date:   Mon Jan 23 03:22:55 2023 +0530

    Merge pull request #12 from ankit-privado/aws-refactoring

    Refactoring and branch name support

commit 22596a823f4cdcd6f6e3ef924d35c6aa41d51588
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Mon Jan 23 03:21:29 2023 +0530

    Refactoring

commit 4d71ac7f3b2fe75fd138c2ba3726ded3fe81fdb9
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Mon Jan 23 03:09:35 2023 +0530

    Refactoring

commit f99344b66dc64348240c8f9730feb02170dde9b5
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Mon Jan 23 01:23:40 2023 +0530

    Refactoring and Branch Name

commit 52d3ef3f97870f960ad72214736a313c0ae227ab
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Fri Jan 20 16:10:39 2023 +0530

    Code Refactoring

commit 9870e8399aefd035e71fa5d143773aa12b38dfa6
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Fri Jan 20 12:53:22 2023 +0530

    Minor change in time

commit cd7f2b8980a03e448563f52739856dff86d9b849
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Fri Jan 20 12:29:18 2023 +0530

    Time issue fixed

commit 4222b7b73b0b06eaef9253625725c93ae0709cd3
Merge: 4d9efbb bff19f0
Author: Ankit Kumar <118803988+ankit-privado@users.noreply.github.com>
Date:   Fri Jan 20 04:03:20 2023 +0530

    Merge pull request #11 from ankit-privado/aws-test

    Aws test

commit bff19f0b321e4a24ab855ae39b5519332080cf13
Merge: acd1259 ebfdfc0
Author: Ankit Kumar <118803988+ankit-privado@users.noreply.github.com>
Date:   Fri Jan 20 04:02:44 2023 +0530

    Merge pull request #10 from ankit-privado/boost-mode

    feature: add boost mode

commit ebfdfc002b25de57596be55fb18e80c8ae23c91e
Author: Abhinav <abhinav.raj@privado.ai>
Date:   Fri Jan 20 00:41:47 2023 +0530

    feature: add boost mode

commit 4d9efbbc741da47f945357e099c45d17e245bba4
Merge: bd45d26 acd1259
Author: Ankit Kumar <118803988+ankit-privado@users.noreply.github.com>
Date:   Fri Jan 20 00:27:33 2023 +0530

    Merge pull request #9 from ankit-privado/aws-test

    Path Analysis change

commit acd12599c853acdaa5508a0a8e57a5bbaa397ef6
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Thu Jan 19 19:44:27 2023 +0530

    Total flow fix

commit bdddfe28d82fa77be5aa513936a2c03f87b33533
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Thu Jan 19 18:11:25 2023 +0530

    Added path count

commit e71f6a0aad22256c67d805739259fb41cee5a87c
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Thu Jan 19 16:32:08 2023 +0530

    Path analysis change

commit bd45d26350492bd437847983567656750dbe73ca
Merge: 1dea6ae 6a9b0ab
Author: Ankit Kumar <118803988+ankit-privado@users.noreply.github.com>
Date:   Thu Jan 19 13:27:30 2023 +0530

    Merge pull request #8 from abhstabs/dev

    Local path changes

commit 6a9b0abd707208fcb9fd6d3ad3ff903f032f5249
Merge: a16c8f8 1dea6ae
Author: Abhinav <abhinav.raj@privado.ai>
Date:   Thu Jan 19 12:52:06 2023 +0530

    Merge branch 'dev' of https://github.com/ankit-privado/standalone-monitoring-stability into dev

commit a16c8f8c0fc2028e6ab5fb1af8980a732a2e7059
Author: Abhinav <abhinav.raj@privado.ai>
Date:   Thu Jan 19 12:52:03 2023 +0530

    feature: add support for local paths

commit 2f1ff9a46168594d6348b9b42cdff56a5dac1d41
Merge: 559a63e 31c4f8a
Author: Abhinav <abhinav.raj@privado.ai>
Date:   Thu Jan 19 11:53:18 2023 +0530

    Merge branch 'main' of https://github.com/ankit-privado/standalone-monitoring-stability into dev

commit 1dea6ae8b9231088175f2d6b268103fab961f547
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Wed Jan 18 22:03:10 2023 +0530

    Path Analysis Change

commit 559a63e2ae789fda80e3a2ee01e00480b3ea423a
Author: Ankit Kumar <ankit.kumar@privado.ai>
Date:   Tue Jan 17 17:53:16 2023 +0530

    Path analysis change

+++ utils/delete.py
@@ -1,16 +1,17 @@
 import os
 import shutil
 
-def delete_action(action):
+def delete_action(action, boost: bool):
 
     if action:
         hard_clean()
     else:
-        soft_clean()
+        soft_clean(boost)
 
 # delete all temp file expect privado-core and privado
-def soft_clean():
-    delete_dir("/temp/binary/")
+def soft_clean(boost: bool):
+    if not boost:
+        delete_dir("/temp/binary/")
     delete_dir("/temp/cpu_mem/")
     delete_dir("/temp/repos")
     delete_dir("/result/")
@@ -25,6 +26,7 @@ def delete_dir(path):
     if os.path.isdir(delete_path):
         shutil.rmtree(delete_path, ignore_errors=False)
 
-def clean_after_scan():
-    delete_dir("/temp/binary/")
+def clean_after_scan(boost: bool):
+    if not boost:
+        delete_dir("/temp/binary/")
     delete_dir("/temp/repos/")
